Move DEFINE_SORT to enums.ts
